requirements.txt
transformers
re
collections
langid
deep_translator
textblob
gensim.models
urllib.request
gzip
os
stanza
tkinter
threading
queue
